requirements.txt
Django==3.2.4
django-extensions==1.9.8
#git+https://github.com/dbca-wa/dpaw-utils.git@0.4.2#egg=dpaw-utils
git+https://github.com/xzzy/webtemplate.git#egg=webtemplate_dbca
git+https://github.com/dbca-wa/dbca-utils@1.1.1#egg=dbca-utils
psycopg2-binary==2.8.5
ipython==7.19.0
django-utils-six==2.0
social-auth-app-django==2.1.0
git+https://github.com/scottp-dpaw/social-core.git@email_fix#egg=social-auth-core
django_bootstrap3==15.0.0
django-reversion==5.0.0
django-reversion-compare==0.15.0
django-cron==0.5.0
djangorestframework==3.12.0
djangorestframework-gis==0.16
djangorestframework-csv==2.0.0
pycountry==17.1.8
geojson==1.3.3
django-confy==1.0.4
gunicorn==19.9.0
dj-static==0.0.6
dj-database-url==0.5.0
whitenoise==5.3.0
django-rest-swagger==2.2.0

# Needed for shell_plus crash issue
jedi==0.17.2
parso==0.7.1

# next packages required for geopandas
# rtree requires 'sudo apt install libspatialindex-dev'
geopandas==0.10.2
shapely==1.8.0
pyproj==3.3.0
pandas==1.3.5
fiona==1.8.20
numpy==1.22.1
python-dateutil==2.8.2
munch==2.5.0
cligj==0.7.2
click-plugins==1.1.1
rtree==0.9.7

# next packages required for geopandas -- plotting
cycler==0.11.0
fonttools==4.28.5
kiwisolver==1.3.2
matplotlib==3.5.1
packaging==21.3
pyparsing==3.0.7


